Fixup/rearrange bloom types
