+++ crates/kitsune_p2p/kitsune_p2p/Cargo.toml
@@ -21,6 +21,7 @@ ghost_actor = "=0.3.0-alpha.6"
 governor = "0.3.2"
 itertools = "0.10"
 kitsune_p2p_fetch = { version = "^0.3.0-beta-dev.7", path = "../fetch" }
+kitsune_p2p_gossip = { version = "^0.1.0", path = "../gossip" }
 kitsune_p2p_mdns = { version = "^0.2.0", path = "../mdns" }
 kitsune_p2p_proxy = { version = "^0.3.0-beta-dev.6", path = "../proxy" }
 kitsune_p2p_timestamp = { version = "^0.3.0-beta-dev.0", path = "../timestamp", features = ["now"] }
